Fixed specular rendering for terrain by using a 32 bit position buffer. Perfomance hit seems small.
Former-commit-id: 8415c88f4aa609f2860d87c23bf951020828bc73
Former-commit-id: bdf2e9420fb5143f227720fb4cb74175e95f7571
Former-commit-id: 20896e1ab9c5bbe7a9066c0c2000b48fc0c4b03e

+++ src/renderer/GBuffer.cpp
@@ -9,10 +9,14 @@ void GBuffer::create(size_t width, size_t height)
 	glGenFramebuffers(1, &g_buffer);
 	glBindFramebuffer(GL_FRAMEBUFFER, g_buffer);
 
-	// High quality position buffer 
+	// Ultra-High quality position buffer 
+	// TODO: Maybe this is overkill, but it's needed for proper 
+	// solar-system sized lighting. If we don't need that, we
+	// could use a GL_RGB16F buffer, but then planets would not be
+	// render properly on a lot of cases.
 	glGenTextures(1, &g_pos);
 	glBindTexture(GL_TEXTURE_2D, g_pos);
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB16F, (int)width, (int)height, 0, GL_RGB, GL_FLOAT, 0);
+	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, (int)width, (int)height, 0, GL_RGB, GL_FLOAT, 0);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 	glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, g_pos, 0);
